polkitd: Add ability to make check methods to return on request
Polkit interactions may require user action to input credentials, this
can lead to delayed replies that can last a potentially infinite time.

We need to be able to test such scenarios, and so add adding mock
methods allowing to perform this, while the CheckAuthorization() method
is now async and able to return on request.